requirements.txt
acrcloud==14.10.2020
aiofiles==23.2.1
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
magic-filter==1.0.12
moviepy==2.1.2
multidict==6.1.0
numpy==2.2.2
pillow==10.4.0
proglog==0.1.10
propcache==0.2.1
psycopg2-binary==2.9.10
pyacrcloud==1.0.8
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.37
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
